requirements.txt
dashscope~=1.20.14
python-dotenv~=1.0.1
pandas~=1.5.3